+++ src/TemplateMethodBundle/Drink/Coffee.php
@@ -7,46 +7,15 @@
 
 namespace TemplateMethodBundle\Drink;
 
-class Coffee
+class Coffee extends CaffeineBeverage
 {
-    /**
-     * @var array
-     */
-    private $debug = [];
-
-    public function prepareRecipe(): void
-    {
-        $this->debug[] = $this->boilWater();
-        $this->debug[] = $this->brewCoffeeGrinds();
-        $this->debug[] = $this->pourInCup();
-        $this->debug[] = $this->addSugarAndMilk();
-    }
-
-    public function boilWater(): string
-    {
-        return "Boiling water.";
-    }
-
-    public function brewCoffeeGrinds(): string
+    public function brew(): string
     {
         return "Dripping Coffee through filter.";
     }
 
-    public function pourInCup(): string
-    {
-        return "Pouring into cup.";
-    }
-
-    public function addSugarAndMilk(): string
+    public function addCondiments(): string
     {
         return "Adding Sugar and Milk";
-    }
-
-    /**
-     * @return array
-     */
-    public function getDebug(): array
-    {
-        return $this->debug;
     }
 }
